+++ src/forze/application/composition/__init__.py
@@ -1,7 +1,8 @@
-from .document import build_document_registry
+from .document import build_document_facade, build_document_registry
 
 # ----------------------- #
 
 __all__ = [
     "build_document_registry",
+    "build_document_facade",
 ]