+++ src/forze/domain/constants.py
@@ -3,6 +3,7 @@
 from typing import Final
 
 # ----------------------- #
+#! Maybe these constants as well ...
 
 ID_FIELD: Final[str] = "id"
 """Document identifier field."""
@@ -16,13 +17,14 @@ SOFT_DELETE_FIELD: Final[str] = "is_deleted"
 NUMBER_ID_FIELD: Final[str] = "number_id"
 """Number identifier field."""
 
-TENANT_ID_FIELD: Final[str] = "tenant_id"  #! Should it be here or in the infra layer ?
-"""Tenant identifier field."""
-
 # ....................... #
+#! Move these constants to the infra layer ! at least not domain definitely
 
 HISTORY_SOURCE_FIELD: Final[str] = "source"
 """History source field."""
 
 HISTORY_DATA_FIELD: Final[str] = "data"
 """History data field."""
+
+TENANT_ID_FIELD: Final[str] = "tenant_id"  #! Should it be here or in the infra layer ?
+"""Tenant identifier field."""