+++ src/forze/infra/providers/redis/gateways/stream.py
@@ -37,7 +37,7 @@ class RedisStreamGateway[M: BaseModel](StreamPort[M]):
     client: RedisClient
     model: type[M]
 
-    # Non initable fields
+    # Non initable fields #! TODO: replace with static definitions
     json_codec: JsonCodec = attrs.field(factory=JsonCodec, init=False)
     text_codec: TextCodec = attrs.field(factory=TextCodec, init=False)